feat: reactivity in Professions
