--- src/lib/proxies/Professions.ts
@@ -1,32 +1,0 @@
-export enum Profession {
-	Rancher = 0,
-	Tiller = 1,
-	Coopmaster = 2,
-	Shepherd = 3,
-	Artisan = 4,
-	Agriculturist = 5,
-	Fisher = 6,
-	Trapper = 7,
-	Angler = 8,
-	Pirate = 9,
-	Mariner = 10,
-	Luremaster = 11,
-	Forester = 12,
-	Gatherer = 13,
-	Lumberjack = 14,
-	Tapper = 15,
-	Botanist = 16,
-	Tracker = 17,
-	Miner = 18,
-	Geologist = 19,
-	Blacksmith = 20,
-	Prospector = 21,
-	Excavator = 22,
-	Gemologist = 23,
-	Fighter = 24,
-	Scout = 25,
-	Brute = 26,
-	Defender = 27,
-	Acrobat = 28,
-	Desperado = 29,
-}
